requirements.txt
Flask
torch
torchvision
Pillow
numpy
pandas
scipy
scikit-learn
matplotlib
seaborn
tqdm
torchsummary
opencv-python-headless 